FEE_test is much less flaky
Presumably as a consequence of JJ's removal of find_baseline from the
chain, or some other change (perhaps in the Py3 work) that reduces the
abundance of NaNs.

Reduced max_runs, min_passes from 10,5 to 3,1

Also reduced the diff cutoff from 0.5 to 0.2

This will speed up the test suite ever so slightly.