Higher-level installation management utilities
This work aims to make it easier for IC users and developers to manage
the installation and preparation of IC code.

The `manage.sh` bash script contains a number of functions which
perform various subsets of the operations needed in various
situations.

The main possibilities are

Install everything and check

    source manage.sh install_and_check 3.5

1. Installs conda if necessary
2. Creates a python 3.5 conda environment with all required dependencies
3. Activates the environment
4. Sets some environment variables needed by the IC code
5. Downloads the test database
6. Compiles IC components written in Cython
7. Runs the test suite

Start working in an existing environment

    source manage.sh work_in_python_version 3.5

As above, skipping the first two steps.

    source manage.sh make_environment 2.7

Creates a new conda environment for specified Python version. To work
in that environment, use `work_in_python_version` as shown above.

Compile recent changes and run the test suite

    bash manage.sh compile_and_test

Download the latest version of the test database.

    bash manage.sh download_test_db