Move tests next to the implementations
The test directory has been removed and any tests that used to be
there were placed in the same directory as the thing that they
test. Any data files used by the tests were moved into a subdirectory
called `test_data` of the directory containing the tests.

The idea is to keep test and implementations together, to encourage
people to read and write the tests, aswell as save them time by not
having to wonder where they are.