Fix #71 : manage.sh usage message on OS X
Correctly identifying whether a shell script is sourced or executed
and extracting the name of the script in a portable and reliable way
across different versions of bash and related shells such as zsh, is
not worth the trouble. This script is not meant to be relocatable; its
whole purpose is to sit in the top directory of IC, so I have taken
the simple and pragmatic approach: hard-wire its name into the usage
message.