Enable concurrent test execution
This uses the `pytest-xdist` extension.

Running the tests concurrently
==============================

`pytest-xdist` allows the concurrent execution of tests on `N` CPUs
by adding the `-n N` option to `pytest`. Alternatively, the number of
available CPUs can be discovered automaticall with `-n auto`.

Unfortunately, some of our tests cannot be run concurrently, as they
form a chain where the outpot of one test serves as the input to
another. Such tests have been marked with `@mark.serial`. With this in
place, the concurrent testing is a two step process:

- `pytest -m "not serial"`
- `pytest -m      serial`

Two new `manage.sh` commands (`run_tests_par` and
`compile_and_test_par`) have been added to this effect.

This solution is not entirely satisfactory, and hopefully something
better can be found.

Given the small number of tests we currently have, and the fact that
about half of the test time is spent in the three, slow, serial tests,
which must be run separately, we don't see any decrease in the overall
testing time: the overhead of setting up the test on multiple
processors is high compared to the time it takes for the concurrent
tests to run.

As the number of tests increases, concurrent testing should become
more worthwhile.

Other xdist features
====================

- `--boxed` runs each test in a separate process, thus preventing a
  segfaulting test from crashing the whole test suite.

- Remote execution: the tests can be sent for execution to remote
  hosts.

- The tests can be made to execute on different versions of python.