Add test suite to document "empty energy plane bug"
In a few cases (less than 0.1 % of the events) NEXT simulation (NEXUS)
produces events with no energy in the PMTs of the energy plane. This
is not strictly a bug (in the sense that it simply reflects a
pathological type of events that occur rarely), but it can produce a
crash in the processing of the data.

While a protection could be set alreay at DIOMIRA level (not accpeting
events with empty energy plane) this would require a rather expensive
test from the computational point of view. A better place to set the
protection is down the processing line, in Irene, asking that the
calibrated sum of the PMTs is not zero.

Function test_diomira_identify_bug() does what its name says. It reads
a file with a single-event of type "empty energy plane", and asserts
that all PMTs are empy. Notice that the test is intended to fail if
it runs in a normal file. The purpose of this test is to demonstrate
that empty "empty energy plane events" exist.

To solve the problem, a protecction is added in irene.py which skip
events where the calibrates (ZS) sum is zero. A second test,
(test_empty_events()) has been
added checking that irene does not fail when reading a one-event file
containing an "empty energy plane" event. The test also checks that
the event is counted.

Two files are needed for the suite. irene_bug_Kr_ACTIVE_7bar_MCRD.h5
is use by test_diomira_identify_bug(), while
irene_bug_Kr_ACTIVE_7bar_RWF.h5 is used by test_empty_events().
Notice that test_empty_events() would fail if the "empty event" slips
through the protection.

Finally, irene.py has been modified so that it prints (optionally) the
number of empty events skipped during a run (test_empty_events()
checks that the number is correct). The corresponding flag to swith
on/off the option has been added to irene.conf