Use `np.histogram` instead of `plt.hist` and fix #95
`test_define_window` was core dumping in headless environments (for
us, this meant Linux-on-Travis). This was caused by the use of
`matplotlib.pyplot.hist` in order to *calculate* a histogram inside
the test. Because `plt.hist` tries to *display* the histogram, it's
not surprising that it crashed on systems that did not have displays.

As the test had no real need to actually display the histogram, merely
to calculate it, this problem was easily solved by using
`numpy.histogram` which performs the same calculation, without the
(pointless and damaging) displaying side-effect.

Moral of the story: don't use drawing functions to perform pure
calculations!