Clean up `.pyx` files
Mostly JCK-1 stuff.

However, most of these files seem to have no tests! (Remember the
rule: if a change to existing code doesn't make an existing test fail,
then that change must be ok.)

There seems to be a lot of copy-n-paste code duplication in these
files, but that's something that will require a bit of thought to
consider (unlike these changes, which were mostly mechanical and
exactly the sort of thing I can do to relax after a long day at the
office.)