Add test suite for configure module.
There was a subtle and hard-to-catch "bug" in configure.py. The bug
itself appears to be in argparse which in python 3 will not take an
int as an argument (as was the case in python2.7). The fix is simply
to modify the argument to str and later cast to int.

A test suite has been added, and a notebook describing the issue is
avaialble at
https://anaconda.org/nexticw/configure_test/notebook

Notice that diomira.ipynb is deleted in this commit. This goes in line
with the policy to hold all notebooks (including those of IC) in ICARO.