Bug in configure module identifed. Test added.
configure.py fails in MacOS Xwhen passing an integer argument through argparse.
This is demonstrated by a test (test_configure), marked as xfail. The
test will fail if run normally but will not interrumpt execution of
tests. Run  pytest --runxfail configure_test.py to force a failure and
report.

Notice that diomira.ipynb is deleted in this commit. This goes in line
with the policy to hold all notebooks (including those of IC) in ICARO.