Add comprehensive test for configuration parser.
This is an expansion of the test that appeared in JJ's (withdrawn) PRs
111 and 112.

Rather than using the real `irene.conf` file (and commenting in the
test that the test will fail if anyone ever changes that file), we
create temporary config files from within the test itself. This allows
us to inject a variety of values into the config file used by the
test, aswell as ensuring that changes to `irene.conf` will not break
the test.

We take care to use different values for all the settings wherever
possible, to minimize the chances of accidentally getting the right
answer.

The test is parametrized, to allow easy orthogonal testing of
independent arguments, and easy addition of new test cases.