Add numpy array creation utilities and their tests
`np_range`, `np_reverse_range` and `np_constant`

The tests are written using hypothesis. The value of these tests is
perhaps greater from the perspective of demonstrating and exploring
hypothesis, rather their value as tests for these almost trivial
functions.

One challenge in writing the tests was the fact that a tiny step
size (as compared to the difference between start and stop) resulted
in the construction of an array with an unreasonable number of
elements, which resulted in all the available memory being eaten up
and the machine grinding to a halt. The solution uses hypothesis'
`composite` decorator to generate step sizes which can be both
positive and negative, but not too close to zero.