Change in calibrated_pmt_sum: propagate to irene and to tests
calibrated_pmt_sum (cython version) returns now a tuple with the calibrated
sum (csum) computed without MAU threshold, and the csum computed with
MAU (previously it returned only the MAU-CSUM). The python version
returns the csum, csum-MAU and the MAU themselves (for the 12 pmts).

The change of interface is propagated to irene and to the various test functions.