A best solution for output file
All working cities need to write down their
data in an output file. The name of the output file and the desired
compression is now part of the base class City.

In addition, Irene opens now the ouput file as a context (using with)
so that the file is automatically closed.