Avoid repetition of namedtuple name
Standard usage of namedtuples violates the DRY (Do not Repeat
Yourself) principle:

    Name = namedtuple('Name', 'a b c d')

requires the repetition of `Name`.

With isolated uses of `namedtuple` it is not worth worrying about this
too much, but in `core/params.py` we create a whole bunch of
`namedtuple`s in succession, and expect more in the future.

So we get rid of the name repetition with a little metaprogramming:
specify the name and the attributes of the `namedtuple` and loop over
these specs to create them and bind them to the module.

Note that the original code already contained a mismatch in the
repeated information for SensorParam(s)!