Demonstrate and fix bug in _store_pmaps
Fix #151

_store_pmaps calls _store_s12 wihout specifying the table that needs
to be stored. The function points always to table s1t (the correct
behaviour is to receive either s1t or s2t by argument and point to
it).

_store_s12 now receives the tables by argument. In addition the tables
are not flushed in the store functions but only at the end of the loop
(flushing every event is not necessary and may hinder performance)

A test has been added, checking that the contents of the tables in
memory (s1t, s2t and s2sit) are the same than the tables in file.