Implement pmaps writer
This work is a complete mess!

The interfaces and other choices made in components brought together
by this work, need serious review.

However, as the functionality is urgently needed, I have hacked
together something that seems to work, and am presenting it here for
immediate inclusion, leaving technical debt reduction until a calmer
time.

This is a single commit resulting from the squashing of numerous steps
by @jjgomezcadenas and @jacg. The more important portions of the
original commit messages follow.

======================================================================

(This class described just below has now been replaced by a context
manager.)

PMAPS are saved in three tables. The operation requires to set the
store file as well as the PMAPS tables before the event loop, save
PMAPS once per event in the event loop and flush the tables before
closing PMAPS store table. A class is a convenient way to handle the
management of tables and hid unncessary details to the client.

The writer has a single public function, that accepts S1, S2 and S2Si
objects ans saves them to pytables.

======================================================================

After discussion (see #165) we have agreed to use always an absolute
event number as key for S12 and S12L. Thus:

S12  = {absolute_evt_number : {peak : namedtuple('t' 'E')}}
S2Si = {absolute_evt_number : {peak : {nsipm:[E]}}}

It follows that the fields evtDaq and nsample are not needed and can
be removed from tables.

======================================================================

Extend writer to add the capability to write run and event info.

When writting PMAPS to file it is necessary to add two additional
tables to the PMAPS tables, one describing run info the other
describing event info. This commit modifies the writer interface so
that run, event and timestamps are saved for each event.

======================================================================

New format for S2Si.

The new format for S2Si required modifications in both the writer and
reader functions. The tests have been modified accordingly and
expanded. The test now writes from transient to file, then reads back
from file and compares with the original objects, providing a robust
consistency check.

======================================================================

Clean up pmaps functions.

module pmaps_functions.py contained a number of functions involving
manipulations of data frames, which are not used in main stream
reconstruction and not formally tested. The functions are removed from
this module and placed in a new module, df_pmaps.py.

======================================================================

Complete refactoring of Irene.

======================================================================

Return S12/S2Si instances from find functions

They used to return dictionaries, forcing the client to make the
conversion.

======================================================================

Add file property to pmap_writer

This is a (hopefully) temporary workaround to get the new
MTTracksWriter to work in the half-refactored Irene.

======================================================================

Move pmaps fixtures up to conftest in invisible_cities

So that they can be reused in the pmap_writer tests

======================================================================

Use pmaps reading fixtures in writer test

======================================================================

Make df_to_dict functions return S12/S2Si instances

======================================================================

Fundamentally, the interfaces are abominable and must be changed. But
we need this functionality in the short-term, so we'll go with this
for now.