Add a PmapWriter class to write PMaps to file.
PMAPS are saved in three tables. The operation requires to set the
store file as well as the PMAPS tables before the event loop, save
PMAPS once per event in the event loop and flush the tables before
closing PMAPS store table. A class is a convenient way to handle the
management of tables and hid unncessary details to the client.

The writer has a single public function, that accepts S1, S2 and S2Si
objects ans saves them to pytables.