Fix #149
The write-only, buggy function `s12df_to_s12l` (the bulk of whose
implementation was in the Cython function `cdf_to_dict`) has been
replaced with a single Cython function `df_to_pmaps_dict`.

An alterative implementation replaces this one in the next commit.

Note: if the order of events in the input dataframe is not
monotonically increasing *and* an event number limit is set, silent
errors are likely to occur.

Extensive tests for the feature have been added.