Introduce interface to convert S2Si DF to dictionaries
A new cython function (df_to_s2si_dict) has been implemented to
provide a better, more comfortable interface with S2Si PMaps. The
function takes the S2Si PMaps in DF format and provides a dictionary
like {evt: {peak: {sipm: E}}}, where evt is the event number, peak is
the peak number, sipm is the SiPM ID number and E is a numpy array
with the charge of the corresponding SiPM for each time slice. These
slices are matched with those of the S2-PMT PMaps, so no time array is
given. In order to fulfill this condition, the E array is filled with
zeros when empty (no charge in the SiPM) slices are found.