Write run info & evt number
Diomira writes run rumber in one table and event number+timestamp in
another table. There is a new option for configuration files (or
command line) to specify which number should be assigned to the first
event in the file.

This is useful for MC productions where we want to keep an absolute
numbering for all the events in one MC run.