Add option to calibrated_pmt_sum
Current version of calibrated_pmt_sum assumes that all the PMTs in the
energy plane are active. This is not necessarily true. This commit
modifies the function to pass the list active_pmt_index which
specifies the indexes of active PMTs (PMT indexes run from 0 to 11).
if the list is empty all 12 PMTs are considered, as before.