Use dictionary instead of seqence of ifs
The signal to noise ratio in the original was pretty low: endless
repetition of

    if name == ????? return tp.Filters(complevel=?, complib=????)

where the ???? represent signal, and the rest is boilerplate.

Dictionaries exist for good reason, and we are all familiar with
them. Let's not re-invent this particular wheel.

It is worrying that no current test spots serious and obvious problems
with this function. For example, try removing the return altogether,
and all our tests pass!