Fix issue 174 + test
There was some mismatching between the number of slices in the cathode
and in the anode. This was caused by a difference in the time value of
consecutive slices smaller than one mus. Although this is still the
case, the number of slices has been forced to be the same in both
planes.