Add funcionality to work with an arbitrary number of PMTs.
So far the software has assume that all PMTs in the detector were
active, so for example the calibrated sum was performed over all the
PMTs in the event. This is not the general case. The new
version of the function takes a list of active PMTs in the argument
and performs the sum over those PMTs. The default behaviour is the
same than before (if no list is specified the sum is taken over all
PMTs).

Several functions have been modified to add the same feature.
Tests have been added to check the new feature

In addition a number of new named tuples have been added to params. They represent
objects that often belowng together:

RawVectors: Raw waveforms, together with the lists of active pmts and sipms
CalPMT: calibration coefficients
PmapParams: groups collections of parameters for PMAPS calculations
CalibratedPMT: calibrated PMTs and Calibrated PMTs with MAU
S1PMaps: S1 objects (play an important role for event selection)

The changes are backward compatible with previous tuple definitions.

Finally a few extra plotting functions have been added.