Add read-only user to DB & Change query to deal with missing PMTs
We now use a read only user to access DB and we use the same database
used by ART. This implies changes in database code and also in some
tests where run number where not specified (for MC so far we use 0).

One important change is that NEXT database assigns negative signs to
the calibration constants, which are positive (as they should!) in
IC. We now return them in absolute value.