Add first version of Dorothea and related cities
Dorothea is a new city that builds a DST with the relevant info for
Kr-like events. The introduction of this city required the coding of
some extra structures:

S12SelectorCity takes the pmaps and selects those peaks satisfying
some configurable conditions (for that, it uses the functions added to
peak_functions). After selections, it builds an instance of the
PointLikeEvent class and returns it, provided that the number of S1
and S2 signals encountered are correct.

Dorothea extends S12SelectorCity and City to run over a list of pmap
files and write those events passing all filters. The persistency is
managed by Kr_writer (dst_io.py) which is a concrete implementation of
the DST_writer class.