Minor fixes
Remove magic number

Cosmetics: wrap and space out monster line

Remove gratuitous semicolons

By making it a subclass of list, you are saying to clients that they
can do to it whatever can be done to lists in general (and more). That
includes

+ sorting
+ reversing
+ clearing
+ popping
+ etc.

I very much doubt that you want all these to be in the interface of
HitCollection!

I also see that HitCollection is a subclass of Event. That also smells
very fishy indeed.