Break monolithic test into digestible pieces
`test_paolina` lumped together a lot of stuff. Specifically, five
different functions from `paolina_functions` were all being tested in
this single test.

Tests should be, if at all possible,

+ small

+ easily understandable by humans

+ orthogonal

In this case, some of the functions being tested depend on the results
produced by some of the others. So splitting the test up naively would
lead to repeated computation of the same results, which could be

+ verbose

+ wasteful, if the functions take non-trivial times to run.

So we create a fixture corresponding to each function, and give the
fixture session scope. In this way, each function will only be
executed once per test session, but its results can be reused
elegantly and cheaply in many different tests.