Decouple find_extrema from shortest path discovery
Originally `find_extrema` calculated the shortest paths between all
pairs of nodes, itself, and kept this information internally. But this
information will be needed to identify which voxels should be included
in a blob.

Rather than making find_extrema return this infomation, (thus making
it responsible for two different things) make it receive this matrix
as input. Whoever calculates this matrix in the first place, can also
pass it to the blob generator.