Add N_PROC parameter to parallel test execution
By default pytest and the *_par functions in manage.sh use all the
available processors. This is (currently) a sub-optimal choice, and
makes doing anything else on the machine in the meanitime a bit
sluggish.

This change makes run_tests_par and compile_and_test_par accept an
optional argument which specifies the number of processors to use.