Mark private attributes with underscore
The interface of this class consist of the constructor and the
__call__ method. While the rest of the methods in the class were
marked with _ to indicate private-to-the-class, the attributes were
not.

The test suite contains tests which go around the public
interface (e.g, they test private attributes that start with _). This
may be not needed if correct tests of the public interface are
implemented. Furthermore, tests act as documentation. By using private
attributes in the test one encourages clients to ignore the interface.

We need to check whether these tests are necessary.