Add genuine command line tests
There are tests in the cities which claim to be command line tests,
but they simulate a pseudo-command line call inside Python, without
actually testing how things work in the UNIX shell.

This is why we didn't spot that the relative imports broke the command
line.

The tests introduced here, really do go outside of Python and use the
UNIX shell.