Make MC event numbers unique
The event numbers are made unique by adding the Python hash of the
filename bas to the event number. We wanted to use 64 ints here, to
minimize the probability of collisions, but it all goes wrong if the
hash exceeds the bounds of 32 bit ints. We have yet to identify where
this problem comes from. For now, we wraph the hash to within the
bounds of 32 bits.