Start removing drivers
The drivers are an unnecessary copy-n-past complication. They can be
replaced by a single classmethod which feeds the configuration
directly to the constructor of the relevant class. To make this work
smoothly, the constructor parameter names must match the config file
parameter names exactly.

While working on this change, we decided that we want a single source
of parameter values. At present, any parameter value might come from
the config file or from various constructor defaults. This will lead
to confusion and subtle errors. So we have started to eliminate the
default values in the constructors ...

... but then the units disappear, so we want units to be stated
explicitly in the config file, which requires the config file parser
to be rewritten to understand units.