Refactoring of code to include new classes in Data Model
The definition of classes Cluster and Hit has been improved, and a new
class "xy" representing an (x,y) point has been added to ic_types. The
code has been refactored to include the new Cluster/Hit interfaces.

Previously Cluster was a namedtuple while Hit was a class quite
lousely defined. In the new definition Hit extends Cluster (a Hit is a
Cluster plus a z position plus an S2 energy). The interfaces of both
objects are now more general and extensive.