Refactoring of module peak_functions.py
Module peak_functions.py contains python functions realated with
peak_finding. Many of the functions exist also in (faster) cython
version. This refactoring:

1) Separates functions in the module into private (marked with
_name_of_function) and public. This convention indicates whether
the function is in the IC API (thus accesible by clients, such as
cities) or not in the API (the typical use then is tests).

2) two functions, select_peaks and select_sy were used exclusively by
the S12 selector (in filters module) and have been moved there.

3) Funciton index_from_s2 was used only by sipm_s2 and has been moved
inside the scope of the sipm_s2 function.

4) Tests have been modified accordingly to reflect changes in the API
and documentation has been added to the module.