Added safety against nans to Waveform.
The tests were failing (as they should) when either the t or E array
in waveform contained a nan. This situation cannot be excluded a
priory and must be foreseen in the construction of Waveform. A test
for nans in the constructor allows to declare the waveform good (if
there are no nans) or not good if a nan is found.

Throwing an exception is not needed or desirable in this case, since a
nan in the waveform does not represent necesarily an exception, it can
be well due to a malformed data packet (hopefully temporary).