Improve S1/S2 interface.
In order to represent S1 and S2 objects, one has to compromise between
the fact that they are structurally identical but functionally
different. For example, one needs an S2 and not an S1 to build an
S2Si.

Defining both S1 and S2 as instances of the same class S12, and using
a simple label to differentiate them is not the best
strategy (although the dynamic nature of python may negate in fact
that we win much by making different types).

Thus, a (private) base class _S12 is defined and public classes S1 and
S2 are derived from _S12. An enumeration type is added for
convenience.

Tests are adapted to reflect changes and test the behaviour of _S12,
S1 and S2