Add s12_selector.conf and fix conf files.
s12_selector.conf is added as part of the configuration tree, in spite
that there is no s12_selector city. It turns out that selectors are a
different species than cities in sofar that they cannot be passed by
inheritance (this would require double inheritance, something we want
to avoid) but at the same time they can be re-used by several
cities. Thus, it make sense to define default parameters for each
selector we introduce.

Several configuration files have been fixed to overwrite parameters
such as input/output file name, which are city-specific (although a
default is inherited from city).